SAM-6D/requirements.txt
addict==2.4.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
antlr4-python3-runtime==4.9.3
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.1.0
beautifulsoup4==4.13.3
blenderproc==2.8.0
blessed==1.20.0
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cmake==3.31.4
colorlog==6.9.0
comm==0.2.2
ConfigArgParse==1.7
contourpy==1.3.0
cycler==0.12.1
dash==3.0.2
decorator==5.2.1
-e git+https://github.com/Charles23R/deep_6dof_tracking.git@9c3ae8d749477bbe822fda33f06bc2b5f09a1d82#egg=deep_6dof_tracking
distinctipy==1.3.4
einops==0.8.1
exceptiongroup==1.2.2
executing==2.2.0
fastjsonschema==2.21.1
filelock==3.17.0
Flask==3.0.3
fonttools==4.56.0
-e git+https://github.com/Charles23R/FoundationPose-private.git@981917cf10703142781a20341a6115fdef8334e0#egg=foundation_pose
freetype-py==2.5.1
frozenlist==1.5.0
fsspec==2025.2.0
fvcore==0.1.5.post20221221
gdown==5.2.0
gorilla-core==0.2.7.8
gpustat==1.0.0
h5py==3.13.0
huggingface-hub==0.29.1
hydra-colorlog==1.2.0
hydra-core==1.3.2
idna==3.10
imageio==2.37.0
imgaug==0.4.0
importlib_metadata==8.6.1
importlib_resources==6.5.2
iniconfig==2.0.0
iopath==0.1.10
ipython==8.18.1
ipywidgets==8.1.5
-e git+https://github.com/Charles23R/SAM-6D-priv.git@cf5c725a8dc32b5b89f6fbeee5a1a7d26ba01e19#egg=ism&subdirectory=SAM-6D/Instance_Segmentation_Model
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
kiwisolver==1.4.7
kornia==0.8.0
kornia_rs==0.1.8
lazy_loader==0.4
lightning-utilities==0.12.0
lit==18.1.8
lmdb==1.6.2
MarkupSafe==3.0.2
matplotlib==3.9.4
matplotlib-inline==0.1.7
mpmath==1.3.0
multidict==6.1.0
narwhals==1.34.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.2.1
ninja==1.11.1.4
numpy==2.0.2
nvdiffrast @ git+https://github.com/NVlabs/nvdiffrast.git@729261dc64c4241ea36efda84fbf532cc8b425b8
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu11==10.2.10.91
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu11==11.7.4.91
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-ml-py==11.495.46
nvidia-nccl-cu11==2.14.3
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu11==11.7.91
nvidia-nvtx-cu12==12.4.127
omegaconf==2.3.0
open3d==0.19.0
opencv-python==4.11.0.86
packaging==24.2
pandas==2.2.3
parso==0.8.4
-e git+https://github.com/Charles23R/SAM-6D-priv.git@cf5c725a8dc32b5b89f6fbeee5a1a7d26ba01e19#egg=pem&subdirectory=SAM-6D/Pose_Estimation_Model
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.6
plotly==6.0.1
pluggy==1.5.0
pointnet2==0.0.0
portalocker==3.1.1
progressbar==2.5
prompt_toolkit==3.0.50
propcache==0.3.0
protobuf==5.29.3
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycocotools==2.0.8
pyglet==2.1.2
Pygments==2.19.1
pynvml==11.5.3
PyOpenGL==3.1.0
pyparsing==3.2.1
pyquaternion==0.9.9
pyrender==0.1.45
PySocks==1.7.1
pytest==8.3.4
python-dateutil==2.9.0.post0
pytorch-lightning==1.8.6
pytz==2025.1
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
retrying==1.3.4
rpds-py==0.24.0
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
safetensors==0.5.2
scikit-image==0.24.0
scikit-learn==1.6.1
scipy==1.13.1
seaborn==0.13.2
segment_anything @ git+https://github.com/facebookresearch/segment-anything.git@dca509fe793f601edb92606367a655c15ac00fdf
shapely==2.0.7
six==1.17.0
soupsieve==2.6
stack-data==0.6.3
sympy==1.13.1
tabulate==0.9.0
tensorboardX==2.6.2.2
termcolor==2.5.0
threadpoolctl==3.6.0
tifffile==2024.8.30
timm==1.0.14
tomli==2.2.1
torch==2.6.0
torchaudio==2.6.0
torchmetrics==0.10.3
torchvision==0.21.0
tqdm==4.67.1
traitlets==5.14.3
transformations==2024.5.24
trimesh==4.0.8
triton==3.2.0
typing_extensions==4.12.2
tzdata==2025.1
ultralytics==8.0.135
urllib3==2.3.0
warp-lang==1.7.0
wcwidth==0.2.13
Werkzeug==3.0.6
widgetsnbextension==4.0.13
yacs==0.1.8
yapf==0.43.0
yarl==1.18.3
zipp==3.21.0
